utils.py
import cv2

def opencv_sketch(image_path):
    img = cv2.imread(image_path)

    gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
    invert = cv2.bitwise_not(gray)
    blur = cv2.GaussianBlur(invert, (21, 21), 0)
    inverted_blur = cv2.bitwise_not(blur)

    sketch = cv2.divide(gray, inverted_blur, scale=256.0)

    return sketch