requirements.txt
streamlit
pillow
numpy
opencv-python-headless
openai